Refactor: 외부 클릭시 닫히는 동작 hook으로 빼기
